Add container for hero section to control width
Added a hero-section container around the hero-container to better
control the width of the hero-container. hero-container now uses
75vw under the 100vw of the hero-section.

Origionally did this to try and eliminate widths on hero-text and
hero-image, but did not like how the site performed when resized.
Ended up keeping the width of each and added flex-shrink: 0 to them.

Going to look into a better way to do this in the future, but for now
it is to spec.