style quote section and switch to link for font
Add styles for quote section and its children. The font specs do not
line up with the results and sizes seem off. Tried to switch from
using import in CSS to links in the head of index.html but it did not
improve. Made the decision to get it as close as I can and move on to
finish the project as a pixel perfect recreation is not the goal of
this particular project.

Fun note, watching pokemon season one in the background while
working <3